# esm/models/bind_resource.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class BindResource(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, app_guid: str=None, route: str=None):
        """
        BindResource - a model defined in Swagger

        :param app_guid: The app_guid of this BindResource.
        :type app_guid: str
        :param route: The route of this BindResource.
        :type route: str
        """
        self.swagger_types = {
            'app_guid': str,
            'route': str
        }

        self.attribute_map = {
            'app_guid': 'app_guid',
            'route': 'route'
        }

        self._app_guid = app_guid
        self._route = route

    @classmethod
    def from_dict(cls, dikt) -> 'BindResource':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BindResource of this BindResource.
        :rtype: BindResource
        """
        return deserialize_model(dikt, cls)

    @property
    def app_guid(self) -> str:
        """
        Gets the app_guid of this BindResource.
        GUID of the application that you want to bind your service to.

        :return: The app_guid of this BindResource.
        :rtype: str
        """
        return self._app_guid

    @app_guid.setter
    def app_guid(self, app_guid: str):
        """
        Sets the app_guid of this BindResource.
        GUID of the application that you want to bind your service to.

        :param app_guid: The app_guid of this BindResource.
        :type app_guid: str
        """

        self._app_guid = app_guid

    @property
    def route(self) -> str:
        """
        Gets the route of this BindResource.
        Route binding

        :return: The route of this BindResource.
        :rtype: str
        """
        return self._route

    @route.setter
    def route(self, route: str):
        """
        Sets the route of this BindResource.
        Route binding

        :param route: The route of this BindResource.
        :type route: str
        """

        self._route = route